refactor(sdk): cut out unnecessary interfaces
Cuts out several unnecessary interfaces that made the docs more
confusing. Also restructures the SDK folder contents a bit to make it a
bit easier to understand.
