+++ .changeset/twelve-ears-explode.md
@@ -1,0 +1,5 @@
+---
+'@eth-optimism/sdk': minor
+---
+
+Refactor SDK folder structure
