+++ packages/sdk/src/bridge-adapters/index.ts
@@ -1,3 +1,4 @@
+export * from './bridge-adapter'
 export * from './standard-bridge'
 export * from './eth-bridge'
 export * from './dai-bridge'
