+++ packages/sdk/src/cross-chain-messenger.ts
@@ -12,7 +12,6 @@ import { sleep, remove0x } from '@eth-optimism/core-utils'
 import { predeploys } from '@eth-optimism/contracts'
 
 import {
-  ICrossChainMessenger,
   OEContracts,
   OEContractsLike,
   MessageLike,
@@ -33,8 +32,7 @@ import {
   BridgeAdapters,
   StateRoot,
   StateRootBatch,
-  IBridgeAdapter,
-} from './interfaces'
+} from './types'
 import {
   toSignerOrProvider,
   toNumber,
@@ -49,14 +47,45 @@ import {
   DEPOSIT_CONFIRMATION_BLOCKS,
   CHAIN_BLOCK_TIMES,
 } from './utils'
+import { IBridgeAdapter } from './bridge-adapters'
 
-export class CrossChainMessenger implements ICrossChainMessenger {
+/**
+ * Handles L1/L2 interactions.
+ */
+export class CrossChainMessenger {
+  /**
+   * Provider connected to the L1 chain.
+   */
   public l1SignerOrProvider: Signer | Provider
+
+  /**
+   * Provider connected to the L2 chain.
+   */
   public l2SignerOrProvider: Signer | Provider
+
+  /**
+   * Chain ID for the L1 network.
+   */
   public l1ChainId: number
+
+  /**
+   * Contract objects attached to their respective providers and addresses.
+   */
   public contracts: OEContracts
+
+  /**
+   * List of custom bridges for the given network.
+   */
   public bridges: BridgeAdapters
+
+  /**
+   * Number of blocks before a deposit is considered confirmed.
+   */
   public depositConfirmationBlocks: number
+
+  /**
+   * Estimated average L1 block time in seconds.
+   */
   public l1BlockTimeSeconds: number
 
   /**
@@ -105,6 +134,9 @@ export class CrossChainMessenger implements ICrossChainMessenger {
     })
   }
 
+  /**
+   * Provider connected to the L1 chain.
+   */
   get l1Provider(): Provider {
     if (Provider.isProvider(this.l1SignerOrProvider)) {
       return this.l1SignerOrProvider
@@ -113,6 +145,9 @@ export class CrossChainMessenger implements ICrossChainMessenger {
     }
   }
 
+  /**
+   * Provider connected to the L2 chain.
+   */
   get l2Provider(): Provider {
     if (Provider.isProvider(this.l2SignerOrProvider)) {
       return this.l2SignerOrProvider
@@ -121,6 +156,9 @@ export class CrossChainMessenger implements ICrossChainMessenger {
     }
   }
 
+  /**
+   * Signer connected to the L1 chain.
+   */
   get l1Signer(): Signer {
     if (Provider.isProvider(this.l1SignerOrProvider)) {
       throw new Error(`messenger has no L1 signer`)
@@ -129,6 +167,9 @@ export class CrossChainMessenger implements ICrossChainMessenger {
     }
   }
 
+  /**
+   * Signer connected to the L2 chain.
+   */
   get l2Signer(): Signer {
     if (Provider.isProvider(this.l2SignerOrProvider)) {
       throw new Error(`messenger has no L2 signer`)
@@ -137,6 +178,16 @@ export class CrossChainMessenger implements ICrossChainMessenger {
     }
   }
 
+  /**
+   * Retrieves all cross chain messages sent within a given transaction.
+   *
+   * @param transaction Transaction hash or receipt to find messages from.
+   * @param opts Options object.
+   * @param opts.direction Direction to search for messages in. If not provided, will attempt to
+   * automatically search both directions under the assumption that a transaction hash will only
+   * exist on one chain. If the hash exists on both chains, will throw an error.
+   * @returns All cross chain messages sent within the transaction.
+   */
   public async getMessagesByTransaction(
     transaction: TransactionLike,
     opts: {
@@ -201,6 +252,19 @@ export class CrossChainMessenger implements ICrossChainMessenger {
       })
   }
 
+  /**
+   * Retrieves all cross chain messages sent by a particular address.
+   *
+   * @param address Address to search for messages from.
+   * @param opts Options object.
+   * @param opts.direction Direction to search for messages in. If not provided, will attempt to
+   * find all messages in both directions.
+   * @param opts.fromBlock Block to start searching for messages from. If not provided, will start
+   * from the first block (block #0).
+   * @param opts.toBlock Block to stop searching for messages at. If not provided, will stop at the
+   * latest known block ("latest").
+   * @returns All cross chain messages sent by the particular address.
+   */
   public async getMessagesByAddress(
     address: AddressLike,
     opts?: {
@@ -218,6 +282,14 @@ export class CrossChainMessenger implements ICrossChainMessenger {
     `)
   }
 
+  /**
+   * Finds the appropriate bridge adapter for a given L1<>L2 token pair. Will throw if no bridges
+   * support the token pair or if more than one bridge supports the token pair.
+   *
+   * @param l1Token L1 token address.
+   * @param l2Token L2 token address.
+   * @returns The appropriate bridge adapter for the given token pair.
+   */
   public async getBridgeForTokenPair(
     l1Token: AddressLike,
     l2Token: AddressLike
@@ -240,6 +312,17 @@ export class CrossChainMessenger implements ICrossChainMessenger {
     return bridges[0]
   }
 
+  /**
+   * Gets all deposits for a given address.
+   *
+   * @param address Address to search for messages from.
+   * @param opts Options object.
+   * @param opts.fromBlock Block to start searching for messages from. If not provided, will start
+   * from the first block (block #0).
+   * @param opts.toBlock Block to stop searching for messages at. If not provided, will stop at the
+   * latest known block ("latest").
+   * @returns All deposit token bridge messages sent by the given address.
+   */
   public async getDepositsByAddress(
     address: AddressLike,
     opts: {
@@ -258,6 +341,17 @@ export class CrossChainMessenger implements ICrossChainMessenger {
     }, [])
   }
 
+  /**
+   * Gets all withdrawals for a given address.
+   *
+   * @param address Address to search for messages from.
+   * @param opts Options object.
+   * @param opts.fromBlock Block to start searching for messages from. If not provided, will start
+   * from the first block (block #0).
+   * @param opts.toBlock Block to stop searching for messages at. If not provided, will stop at the
+   * latest known block ("latest").
+   * @returns All withdrawal token bridge messages sent by the given address.
+   */
   public async getWithdrawalsByAddress(
     address: AddressLike,
     opts: {
@@ -276,6 +370,16 @@ export class CrossChainMessenger implements ICrossChainMessenger {
     }, [])
   }
 
+  /**
+   * Resolves a MessageLike into a CrossChainMessage object.
+   * Unlike other coercion functions, this function is stateful and requires making additional
+   * requests. For now I'm going to keep this function here, but we could consider putting a
+   * similar function inside of utils/coercion.ts if people want to use this without having to
+   * create an entire CrossChainProvider object.
+   *
+   * @param message MessageLike to resolve into a CrossChainMessage.
+   * @returns Message coerced into a CrossChainMessage.
+   */
   public async toCrossChainMessage(
     message: MessageLike
   ): Promise<CrossChainMessage> {
@@ -329,6 +433,12 @@ export class CrossChainMessenger implements ICrossChainMessenger {
     }
   }
 
+  /**
+   * Retrieves the status of a particular message as an enum.
+   *
+   * @param message Cross chain message to check the status of.
+   * @returns Status of the message.
+   */
   public async getMessageStatus(message: MessageLike): Promise<MessageStatus> {
     const resolved = await this.toCrossChainMessage(message)
     const receipt = await this.getMessageReceipt(resolved)
@@ -370,6 +480,13 @@ export class CrossChainMessenger implements ICrossChainMessenger {
     }
   }
 
+  /**
+   * Finds the receipt of the transaction that executed a particular cross chain message.
+   *
+   * @param message Message to find the receipt of.
+   * @returns CrossChainMessage receipt including receipt of the transaction that relayed the
+   * given message.
+   */
   public async getMessageReceipt(
     message: MessageLike
   ): Promise<MessageReceipt> {
@@ -428,6 +545,18 @@ export class CrossChainMessenger implements ICrossChainMessenger {
     return null
   }
 
+  /**
+   * Waits for a message to be executed and returns the receipt of the transaction that executed
+   * the given message.
+   *
+   * @param message Message to wait for.
+   * @param opts Options to pass to the waiting function.
+   * @param opts.confirmations Number of transaction confirmations to wait for before returning.
+   * @param opts.pollIntervalMs Number of milliseconds to wait between polling for the receipt.
+   * @param opts.timeoutMs Milliseconds to wait before timing out.
+   * @returns CrossChainMessage receipt including receipt of the transaction that relayed the
+   * given message.
+   */
   public async waitForMessageReceipt(
     message: MessageLike,
     opts: {
@@ -451,6 +580,16 @@ export class CrossChainMessenger implements ICrossChainMessenger {
     throw new Error(`timed out waiting for message receipt`)
   }
 
+  /**
+   * Estimates the amount of gas required to fully execute a given message on L2. Only applies to
+   * L1 => L2 messages. You would supply this gas limit when sending the message to L2.
+   *
+   * @param message Message get a gas estimate for.
+   * @param opts Options object.
+   * @param opts.bufferPercent Percentage of gas to add to the estimate. Defaults to 20.
+   * @param opts.from Address to use as the sender.
+   * @returns Estimates L2 gas limit.
+   */
   public async estimateL2MessageGasLimit(
     message: MessageRequestLike,
     opts?: {
@@ -484,6 +623,15 @@ export class CrossChainMessenger implements ICrossChainMessenger {
     return estimate.mul(100 + bufferPercent).div(100)
   }
 
+  /**
+   * Returns the estimated amount of time before the message can be executed. When this is a
+   * message being sent to L1, this will return the estimated time until the message will complete
+   * its challenge period. When this is a message being sent to L2, this will return the estimated
+   * amount of time until the message will be picked up and executed on L2.
+   *
+   * @param message Message to estimate the time remaining for.
+   * @returns Estimated amount of time remaining (in seconds) before the message can be executed.
+   */
   public async estimateMessageWaitTimeSeconds(
     message: MessageLike
   ): Promise<number> {
@@ -543,12 +691,25 @@ export class CrossChainMessenger implements ICrossChainMessenger {
     }
   }
 
+  /**
+   * Queries the current challenge period in seconds from the StateCommitmentChain.
+   *
+   * @returns Current challenge period in seconds.
+   */
   public async getChallengePeriodSeconds(): Promise<number> {
     const challengePeriod =
       await this.contracts.l1.StateCommitmentChain.FRAUD_PROOF_WINDOW()
     return challengePeriod.toNumber()
   }
 
+  /**
+   * Returns the state root that corresponds to a given message. This is the state root for the
+   * block in which the transaction was included, as published to the StateCommitmentChain. If the
+   * state root for the given message has not been published yet, this function returns null.
+   *
+   * @param message Message to find a state root for.
+   * @returns State root for the block in which the message was created.
+   */
   public async getMessageStateRoot(
     message: MessageLike
   ): Promise<StateRoot | null> {
@@ -600,6 +761,13 @@ export class CrossChainMessenger implements ICrossChainMessenger {
     }
   }
 
+  /**
+   * Returns the StateBatchAppended event that was emitted when the batch with a given index was
+   * created. Returns null if no such event exists (the batch has not been submitted).
+   *
+   * @param batchIndex Index of the batch to find an event for.
+   * @returns StateBatchAppended event for the batch, or null if no such batch exists.
+   */
   public async getStateBatchAppendedEventByBatchIndex(
     batchIndex: number
   ): Promise<ethers.Event | null> {
@@ -619,6 +787,13 @@ export class CrossChainMessenger implements ICrossChainMessenger {
     }
   }
 
+  /**
+   * Returns the StateBatchAppended event for the batch that includes the transaction with the
+   * given index. Returns null if no such event exists.
+   *
+   * @param transactionIndex Index of the L2 transaction to find an event for.
+   * @returns StateBatchAppended event for the batch that includes the given transaction by index.
+   */
   public async getStateBatchAppendedEventByTransactionIndex(
     transactionIndex: number
   ): Promise<ethers.Event | null> {
@@ -673,6 +848,13 @@ export class CrossChainMessenger implements ICrossChainMessenger {
     return batchEvent
   }
 
+  /**
+   * Returns information about the state root batch that included the state root for the given
+   * transaction by index. Returns null if no such state root has been published yet.
+   *
+   * @param transactionIndex Index of the L2 transaction to find a state root batch for.
+   * @returns State root batch for the given transaction index, or null if none exists yet.
+   */
   public async getStateRootBatchByTransactionIndex(
     transactionIndex: number
   ): Promise<StateRootBatch | null> {
@@ -702,6 +884,12 @@ export class CrossChainMessenger implements ICrossChainMessenger {
     }
   }
 
+  /**
+   * Generates the proof required to finalize an L2 to L1 message.
+   *
+   * @param message Message to generate a proof for.
+   * @returns Proof that can be used to finalize the message.
+   */
   public async getMessageProof(
     message: MessageLike
   ): Promise<CrossChainMessageProof> {
@@ -750,6 +938,17 @@ export class CrossChainMessenger implements ICrossChainMessenger {
     }
   }
 
+  /**
+   * Sends a given cross chain message. Where the message is sent depends on the direction attached
+   * to the message itself.
+   *
+   * @param message Cross chain message to send.
+   * @param opts Additional options.
+   * @param opts.signer Optional signer to use to send the transaction.
+   * @param opts.l2GasLimit Optional gas limit to use for the transaction on L2.
+   * @param opts.overrides Optional transaction overrides.
+   * @returns Transaction response for the message sending transaction.
+   */
   public async sendMessage(
     message: CrossChainMessageRequest,
     opts?: {
@@ -766,6 +965,17 @@ export class CrossChainMessenger implements ICrossChainMessenger {
     }
   }
 
+  /**
+   * Resends a given cross chain message with a different gas limit. Only applies to L1 to L2
+   * messages. If provided an L2 to L1 message, this function will throw an error.
+   *
+   * @param message Cross chain message to resend.
+   * @param messageGasLimit New gas limit to use for the message.
+   * @param opts Additional options.
+   * @param opts.signer Optional signer to use to send the transaction.
+   * @param opts.overrides Optional transaction overrides.
+   * @returns Transaction response for the message resending transaction.
+   */
   public async resendMessage(
     message: MessageLike,
     messageGasLimit: NumberLike,
@@ -783,6 +993,16 @@ export class CrossChainMessenger implements ICrossChainMessenger {
     )
   }
 
+  /**
+   * Finalizes a cross chain message that was sent from L2 to L1. Only applicable for L2 to L1
+   * messages. Will throw an error if the message has not completed its challenge period yet.
+   *
+   * @param message Message to finalize.
+   * @param opts Additional options.
+   * @param opts.signer Optional signer to use to send the transaction.
+   * @param opts.overrides Optional transaction overrides.
+   * @returns Transaction response for the finalization transaction.
+   */
   public async finalizeMessage(
     message: MessageLike,
     opts?: {
@@ -795,6 +1015,17 @@ export class CrossChainMessenger implements ICrossChainMessenger {
     )
   }
 
+  /**
+   * Deposits some ETH into the L2 chain.
+   *
+   * @param amount Amount of ETH to deposit (in wei).
+   * @param opts Additional options.
+   * @param opts.signer Optional signer to use to send the transaction.
+   * @param opts.recipient Optional address to receive the funds on L2. Defaults to sender.
+   * @param opts.l2GasLimit Optional gas limit to use for the transaction on L2.
+   * @param opts.overrides Optional transaction overrides.
+   * @returns Transaction response for the deposit transaction.
+   */
   public async depositETH(
     amount: NumberLike,
     opts?: {
@@ -809,6 +1040,16 @@ export class CrossChainMessenger implements ICrossChainMessenger {
     )
   }
 
+  /**
+   * Withdraws some ETH back to the L1 chain.
+   *
+   * @param amount Amount of ETH to withdraw.
+   * @param opts Additional options.
+   * @param opts.signer Optional signer to use to send the transaction.
+   * @param opts.recipient Optional address to receive the funds on L1. Defaults to sender.
+   * @param opts.overrides Optional transaction overrides.
+   * @returns Transaction response for the withdraw transaction.
+   */
   public async withdrawETH(
     amount: NumberLike,
     opts?: {
@@ -822,6 +1063,19 @@ export class CrossChainMessenger implements ICrossChainMessenger {
     )
   }
 
+  /**
+   * Deposits some ERC20 tokens into the L2 chain.
+   *
+   * @param l1Token Address of the L1 token.
+   * @param l2Token Address of the L2 token.
+   * @param amount Amount to deposit.
+   * @param opts Additional options.
+   * @param opts.signer Optional signer to use to send the transaction.
+   * @param opts.recipient Optional address to receive the funds on L2. Defaults to sender.
+   * @param opts.l2GasLimit Optional gas limit to use for the transaction on L2.
+   * @param opts.overrides Optional transaction overrides.
+   * @returns Transaction response for the deposit transaction.
+   */
   public async depositERC20(
     l1Token: AddressLike,
     l2Token: AddressLike,
@@ -843,6 +1097,18 @@ export class CrossChainMessenger implements ICrossChainMessenger {
     )
   }
 
+  /**
+   * Withdraws some ERC20 tokens back to the L1 chain.
+   *
+   * @param l1Token Address of the L1 token.
+   * @param l2Token Address of the L2 token.
+   * @param amount Amount to withdraw.
+   * @param opts Additional options.
+   * @param opts.signer Optional signer to use to send the transaction.
+   * @param opts.recipient Optional address to receive the funds on L1. Defaults to sender.
+   * @param opts.overrides Optional transaction overrides.
+   * @returns Transaction response for the withdraw transaction.
+   */
   public async withdrawERC20(
     l1Token: AddressLike,
     l2Token: AddressLike,
@@ -863,7 +1129,21 @@ export class CrossChainMessenger implements ICrossChainMessenger {
     )
   }
 
+  /**
+   * Object that holds the functions that generate transactions to be signed by the user.
+   * Follows the pattern used by ethers.js.
+   */
   populateTransaction = {
+    /**
+     * Generates a transaction that sends a given cross chain message. This transaction can be signed
+     * and executed by a signer.
+     *
+     * @param message Cross chain message to send.
+     * @param opts Additional options.
+     * @param opts.l2GasLimit Optional gas limit to use for the transaction on L2.
+     * @param opts.overrides Optional transaction overrides.
+     * @returns Transaction that can be signed and executed to send the message.
+     */
     sendMessage: async (
       message: CrossChainMessageRequest,
       opts?: {
@@ -888,6 +1168,16 @@ export class CrossChainMessenger implements ICrossChainMessenger {
       }
     },
 
+    /**
+     * Generates a transaction that resends a given cross chain message. Only applies to L1 to L2
+     * messages. This transaction can be signed and executed by a signer.
+     *
+     * @param message Cross chain message to resend.
+     * @param messageGasLimit New gas limit to use for the message.
+     * @param opts Additional options.
+     * @param opts.overrides Optional transaction overrides.
+     * @returns Transaction that can be signed and executed to resend the message.
+     */
     resendMessage: async (
       message: MessageLike,
       messageGasLimit: NumberLike,
@@ -911,6 +1201,16 @@ export class CrossChainMessenger implements ICrossChainMessenger {
       )
     },
 
+    /**
+     * Generates a message finalization transaction that can be signed and executed. Only
+     * applicable for L2 to L1 messages. Will throw an error if the message has not completed
+     * its challenge period yet.
+     *
+     * @param message Message to generate the finalization transaction for.
+     * @param opts Additional options.
+     * @param opts.overrides Optional transaction overrides.
+     * @returns Transaction that can be signed and executed to finalize the message.
+     */
     finalizeMessage: async (
       message: MessageLike,
       opts?: {
@@ -933,6 +1233,16 @@ export class CrossChainMessenger implements ICrossChainMessenger {
       )
     },
 
+    /**
+     * Generates a transaction for depositing some ETH into the L2 chain.
+     *
+     * @param amount Amount of ETH to deposit.
+     * @param opts Additional options.
+     * @param opts.recipient Optional address to receive the funds on L2. Defaults to sender.
+     * @param opts.l2GasLimit Optional gas limit to use for the transaction on L2.
+     * @param opts.overrides Optional transaction overrides.
+     * @returns Transaction that can be signed and executed to deposit the ETH.
+     */
     depositETH: async (
       amount: NumberLike,
       opts?: {
@@ -949,6 +1259,15 @@ export class CrossChainMessenger implements ICrossChainMessenger {
       )
     },
 
+    /**
+     * Generates a transaction for withdrawing some ETH back to the L1 chain.
+     *
+     * @param amount Amount of ETH to withdraw.
+     * @param opts Additional options.
+     * @param opts.recipient Optional address to receive the funds on L1. Defaults to sender.
+     * @param opts.overrides Optional transaction overrides.
+     * @returns Transaction that can be signed and executed to withdraw the ETH.
+     */
     withdrawETH: async (
       amount: NumberLike,
       opts?: {
@@ -964,6 +1283,18 @@ export class CrossChainMessenger implements ICrossChainMessenger {
       )
     },
 
+    /**
+     * Generates a transaction for depositing some ERC20 tokens into the L2 chain.
+     *
+     * @param l1Token Address of the L1 token.
+     * @param l2Token Address of the L2 token.
+     * @param amount Amount to deposit.
+     * @param opts Additional options.
+     * @param opts.recipient Optional address to receive the funds on L2. Defaults to sender.
+     * @param opts.l2GasLimit Optional gas limit to use for the transaction on L2.
+     * @param opts.overrides Optional transaction overrides.
+     * @returns Transaction that can be signed and executed to deposit the tokens.
+     */
     depositERC20: async (
       l1Token: AddressLike,
       l2Token: AddressLike,
@@ -978,6 +1309,17 @@ export class CrossChainMessenger implements ICrossChainMessenger {
       return bridge.populateTransaction.deposit(l1Token, l2Token, amount, opts)
     },
 
+    /**
+     * Generates a transaction for withdrawing some ERC20 tokens back to the L1 chain.
+     *
+     * @param l1Token Address of the L1 token.
+     * @param l2Token Address of the L2 token.
+     * @param amount Amount to withdraw.
+     * @param opts Additional options.
+     * @param opts.recipient Optional address to receive the funds on L1. Defaults to sender.
+     * @param opts.overrides Optional transaction overrides.
+     * @returns Transaction that can be signed and executed to withdraw the tokens.
+     */
     withdrawERC20: async (
       l1Token: AddressLike,
       l2Token: AddressLike,
@@ -992,7 +1334,20 @@ export class CrossChainMessenger implements ICrossChainMessenger {
     },
   }
 
+  /**
+   * Object that holds the functions that estimates the gas required for a given transaction.
+   * Follows the pattern used by ethers.js.
+   */
   estimateGas = {
+    /**
+     * Estimates gas required to send a cross chain message.
+     *
+     * @param message Cross chain message to send.
+     * @param opts Additional options.
+     * @param opts.l2GasLimit Optional gas limit to use for the transaction on L2.
+     * @param opts.overrides Optional transaction overrides.
+     * @returns Gas estimate for the transaction.
+     */
     sendMessage: async (
       message: CrossChainMessageRequest,
       opts?: {
@@ -1008,6 +1363,15 @@ export class CrossChainMessenger implements ICrossChainMessenger {
       }
     },
 
+    /**
+     * Estimates gas required to resend a cross chain message. Only applies to L1 to L2 messages.
+     *
+     * @param message Cross chain message to resend.
+     * @param messageGasLimit New gas limit to use for the message.
+     * @param opts Additional options.
+     * @param opts.overrides Optional transaction overrides.
+     * @returns Gas estimate for the transaction.
+     */
     resendMessage: async (
       message: MessageLike,
       messageGasLimit: NumberLike,
@@ -1024,6 +1388,14 @@ export class CrossChainMessenger implements ICrossChainMessenger {
       )
     },
 
+    /**
+     * Estimates gas required to finalize a cross chain message. Only applies to L2 to L1 messages.
+     *
+     * @param message Message to generate the finalization transaction for.
+     * @param opts Additional options.
+     * @param opts.overrides Optional transaction overrides.
+     * @returns Gas estimate for the transaction.
+     */
     finalizeMessage: async (
       message: MessageLike,
       opts?: {
@@ -1035,6 +1407,16 @@ export class CrossChainMessenger implements ICrossChainMessenger {
       )
     },
 
+    /**
+     * Estimates gas required to deposit some ETH into the L2 chain.
+     *
+     * @param amount Amount of ETH to deposit.
+     * @param opts Additional options.
+     * @param opts.recipient Optional address to receive the funds on L2. Defaults to sender.
+     * @param opts.l2GasLimit Optional gas limit to use for the transaction on L2.
+     * @param opts.overrides Optional transaction overrides.
+     * @returns Gas estimate for the transaction.
+     */
     depositETH: async (
       amount: NumberLike,
       opts?: {
@@ -1048,6 +1430,15 @@ export class CrossChainMessenger implements ICrossChainMessenger {
       )
     },
 
+    /**
+     * Estimates gas required to withdraw some ETH back to the L1 chain.
+     *
+     * @param amount Amount of ETH to withdraw.
+     * @param opts Additional options.
+     * @param opts.recipient Optional address to receive the funds on L1. Defaults to sender.
+     * @param opts.overrides Optional transaction overrides.
+     * @returns Gas estimate for the transaction.
+     */
     withdrawETH: async (
       amount: NumberLike,
       opts?: {
@@ -1060,6 +1451,18 @@ export class CrossChainMessenger implements ICrossChainMessenger {
       )
     },
 
+    /**
+     * Estimates gas required to deposit some ERC20 tokens into the L2 chain.
+     *
+     * @param l1Token Address of the L1 token.
+     * @param l2Token Address of the L2 token.
+     * @param amount Amount to deposit.
+     * @param opts Additional options.
+     * @param opts.recipient Optional address to receive the funds on L2. Defaults to sender.
+     * @param opts.l2GasLimit Optional gas limit to use for the transaction on L2.
+     * @param opts.overrides Optional transaction overrides.
+     * @returns Gas estimate for the transaction.
+     */
     depositERC20: async (
       l1Token: AddressLike,
       l2Token: AddressLike,
@@ -1080,6 +1483,17 @@ export class CrossChainMessenger implements ICrossChainMessenger {
       )
     },
 
+    /**
+     * Estimates gas required to withdraw some ERC20 tokens back to the L1 chain.
+     *
+     * @param l1Token Address of the L1 token.
+     * @param l2Token Address of the L2 token.
+     * @param amount Amount to withdraw.
+     * @param opts Additional options.
+     * @param opts.recipient Optional address to receive the funds on L1. Defaults to sender.
+     * @param opts.overrides Optional transaction overrides.
+     * @returns Gas estimate for the transaction.
+     */
     withdrawERC20: async (
       l1Token: AddressLike,
       l2Token: AddressLike,
