+++ packages/sdk/src/index.ts
@@ -1,5 +1,5 @@
-export * from './interfaces'
+export * from './bridge-adapters'
 export * from './utils'
 export * from './cross-chain-messenger'
-export * from './adapters'
 export * from './l2-provider'
+export * from './types'
