--- packages/sdk/src/interfaces/index.ts
@@ -1,4 +1,0 @@
-export * from './bridge-adapter'
-export * from './cross-chain-messenger'
-export * from './l2-provider'
-export * from './types'
