+++ packages/sdk/src/types.ts
@@ -6,8 +6,8 @@ import {
 import { Signer } from '@ethersproject/abstract-signer'
 import { Contract, BigNumber } from 'ethers'
 
-import { ICrossChainMessenger } from './cross-chain-messenger'
-import { IBridgeAdapter } from './bridge-adapter'
+import { CrossChainMessenger } from './cross-chain-messenger'
+import { IBridgeAdapter } from './bridge-adapters/bridge-adapter'
 
 /**
  * L1 contract references.
@@ -76,7 +76,7 @@ export interface OEContractsLike {
 export interface BridgeAdapterData {
   [name: string]: {
     Adapter: new (opts: {
-      messenger: ICrossChainMessenger
+      messenger: CrossChainMessenger
       l1Bridge: AddressLike
       l2Bridge: AddressLike
     }) => IBridgeAdapter
