+++ packages/sdk/src/utils/index.ts
@@ -1,4 +1,4 @@
-export * from './coercion'
+export * from './type-coercion'
 export * from './contracts'
 export * from './message-encoding'
 export * from './type-utils'
